article_webpage/requirements.txt
flask
kafka-python
jinja2